move callable to /styles
