+++ packages/react-bindings/src/index.ts
@@ -13,6 +13,5 @@ export { default as useAccessibility } from './hooks/useAccessibility'
 export { default as unstable_useDispatchEffect } from './hooks/useDispatchEffect'
 export { default as useStateManager } from './hooks/useStateManager'
 
-export { default as callable } from './utils/callable'
 export { default as getElementType } from './utils/getElementType'
 export { default as getUnhandledProps } from './utils/getUnhandledProps'
