+++ packages/react/src/themes/teams/components/Icon/iconStyles.ts
@@ -1,6 +1,6 @@
 import * as _ from 'lodash'
-import { callable } from '@fluentui/react-bindings'
 import {
+  callable,
   ComponentSlotStylesPrepared,
   FontIconSpec,
   ICSSInJSStyle,
